feat(Book): component structure
